Pointer ABI for Windows (#16)
Windows ABI expects the return value to be passed by reference as a first argument. After a bit of experimentation we've settled on using the same Tuple{VecElement} trick as on unix (to ensure 128-bit alignment) and then defining ccall macro that handles the conversion correctly.

A few other small fixes:
- Use correct limb size for MPFR
- Use correct quadoplib name for Win32

I've also had to disable 0.7 support as it doesn't like the splatting syntax we need to use for sprint, but this doesn't seem like a big deal.